+++ frontEnd/package.json
@@ -17,7 +17,8 @@
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.18.0",
     "socket.io-client": "^4.7.2",
-    "uuid": "^9.0.1"
+    "uuid": "^9.0.1",
+    "yjs": "^13.6.8"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.4",
